[FIX] runbot: properly redirect stderr during git archive
In a91b08897fcc stderr content was used to decide if a failure occured
during git archive. The problem is that stderr was not redirected to
subprocess.PIPE and thus lost in the limbo.

With this commit, the return code of each subprocess is used to decide
if an error occured. The first process checked is git because if it
fails, tar will fail too.

Also, each stderr is correctly redirected to subprocess.PIPE.